pyproject.toml
[tool.poe.tasks]
# run with eg `uv run poe fmt`
fmt = "ruff format"
lint = "ruff check --fix"
check = "pyright"
test = "pytest -m 'not integration'"  # fast unit tests only (no Stockfish required)
"test:all" = "pytest"  # all tests including integration
"test:integration" = "pytest -m integration"  # integration tests only
# run all the above
all = [ {ref="fmt"}, {ref="lint"}, {ref="check"}, {ref="test"} ]

"ci:fmt" = "ruff format --check"  # fail if not formatted
"ci:lint" = "ruff check"

[project]
name = "chess_sandbox"
version = "0.1.0"
description = "Chess commentary generation using LLMs and Stockfish"
readme = "README.md"
authors = [
    { name = "Guillaume Allain", email = "guillaume.allain@gmail.com" }
]
keywords = ["chess", "stockfish", "llm", "openai", "claude", "commentary", "analysis"]

# Public libraries should be more lenient
# Internal stuff should enforce ~=3.13!
requires-python = ">=3.13"
dependencies = [
    "chess>=1.11.2",
    "click>=8.3.0",
    "openai>=2.6.1",
    "pydantic>=2.10.4",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "poethepoet>=0.32.0",
    "pyright>=1.1.391",
    "pytest>=8.3.4",
    "ruff>=0.8.5",
]

[tool.uv]
# tell uv we're building a package
# (i.e. something we can distribute for others to use)
# uncomment this to use project.scripts
# package = true

# [project.scripts]
# running `postmodern` will run the `postmodern.main` function
# postmodern = "postmodern.hello:main"

[tool.ruff]
# if this is a library, enter the _minimum_ version you
# want to support, otherwise do py313
target-version = "py313"
line-length = 120  # use whatever number makes you happy

[tool.ruff.lint]
# you can see the looong list of rules here:
# https://docs.astral.sh/ruff/rules/
# here's a couple to start with
select = [
  "A",    # warn about shadowing built-ins
  "E",    # style stuff, whitespaces
  "F",    # important pyflakes lints
  "I",    # import sorting
  "N",    # naming
  "T100", # breakpoints (probably don't want these in prod!)
]
# if you're feeling confident you can do:
# select = ["ALL"]
# and then manually ignore annoying ones:
# ignore = [...]

[tool.ruff.lint.isort]
# so it knows to group first-party stuff last
known-first-party = ["chess_sandbox"]

[tool.pyright]
venvPath = "."          # uv installs the venv in the current dir
venv = ".venv"          #       in a folder called `.venv`
strict = ["**/*.py"]    # use 'strict' checking on all files
pythonVersion = "3.13"  # if library, specify the _lowest_ you support

[tool.pytest.ini_options]
addopts = "--doctest-modules"
markers = [
    "integration: marks tests as integration tests requiring Stockfish (deselect with '-m \"not integration\"')",
]
